services/ai-worker/requirements.txt
pytest
pytest-mock
requests
boto3
pika
pymongo==4.8.0
mutagen==1.47.0
minio
google-generativeai
